requirements.txt
python-telegram-bot>=22.0,<23.0
tzdata>=2024.1
